ISR IRQ refactor
